Added unit tests for the game_message table
* Made the auto increment id field to bigserial instead of serial
* Made create and fetch sql statements for the messages
* messages can be fetched with game_id or game_room_id
* messages can be fetched using offset pagination

+++ database/sqlc/models.go
@@ -30,7 +30,7 @@ type Game struct {
 type GameMessage struct {
 	GameID     uuid.UUID       `json:"game_id"`
 	GameRoomID uuid.UUID       `json:"game_room_id"`
-	MessageID  int32           `json:"message_id"`
+	MessageID  int64           `json:"message_id"`
 	Message    json.RawMessage `json:"message"`
 	CreatedAt  time.Time       `json:"created_at"`
 }
